requirements.txt
gradio
langchain
langchain-core
langchain-google-genai
langgraph
pandas
openpyxl
